refactor: rename src/new/ to src/api-contract/ in both http clients
