+++ packages/app/frontend-http-client/src/api-contract/UnexpectedResponseError.ts
