+++ packages/app/frontend-http-client/src/index.ts
@@ -11,7 +11,7 @@ export {
   sendPut,
   UNKNOWN_SCHEMA,
 } from './client.ts'
-export { type ContractRequestOptions, sendByApiContract } from './new/sendByApiContract.ts'
-export { UnexpectedResponseError } from './new/UnexpectedResponseError.ts'
+export { type ContractRequestOptions, sendByApiContract } from './api-contract/sendByApiContract.ts'
+export { UnexpectedResponseError } from './api-contract/UnexpectedResponseError.ts'
 export type { SseCallbacks, SseConnection, SseRouteRequestParams } from './sse.ts'
 export { connectSseByContract } from './sse.ts'
